WIP on android module updates
